code/requirements.txt
functions-framework==3.*
google-cloud-billing>=1.11
base64
os
json
logging